- Write 24 JUnit4 test - Write new method setNumberText

+++ app/src/main/java/com/example/ilafedoseev/calculatorkotlin/Calculate.kt
@@ -40,7 +40,7 @@ class Calculate {
                     } else if (token.toDouble() > currentNumber) {
                         currentNumber = currentNumber.div(token.toDouble())
                     } else {
-                        currentNumber /= token.toInt()
+                        currentNumber /= token.toDouble()
                     }
                 }
